+++ src/actions/api/backups.js
@@ -25,3 +25,16 @@ export function takeBackup(id) {
     dispatch({ type: UPDATE_BACKUP, backup: json, linodes: id });
   };
 }
+
+export function restoreBackup(linodeId, targetLinode, backupId, overwrite = false) {
+  return async (dispatch, getState) => {
+    const state = getState();
+    const { token } = state.authentication;
+    const response = await fetch(token,
+      `/linodes/${linodeId}/backups/${backupId}/restore`, {
+        method: 'POST',
+        body: JSON.stringify({ linode: targetLinode, overwrite }),
+      });
+    return await response.json();
+  };
+}
